GDB-11134 fix error handling in my settings view (#1676)
* Extracted the ChangeUserPasswordSettingsCtrl as UserSettingsController in separate controller file.
Extracted constants and utilities in separate files.

* Move switching the theme in the success handler after user's data have been updated to prevent the inconsistency in the UI state when a server error occurs and the user update operation fails.
Also reorganized and cleaned the code a bit for better readability and maintainability.

* Fixed collisions after rebase and some more cleanup

* Fixed jsdoc for the showCookiePolicyLink flag. Refactored get principal function.

(cherry picked from commit 46544a82bd2fb449de573907f27b78b4f3fc0fe5)
