+++ src/js/angular/security/services/constants.js
@@ -1,0 +1,5 @@
+export const SYSTEM_REPO = 'SYSTEM';
+export const READ_REPO = 'READ_REPO';
+export const READ_REPO_PREFIX = 'READ_REPO_';
+export const WRITE_REPO = 'WRITE_REPO';
+export const WRITE_REPO_PREFIX = 'WRITE_REPO_';
